first (untested) implementation
